feat(ask-core): F0 slice 1 — extract the shared embedder into @agentskit/ask-core (RFC-0007)
New private package @agentskit/ask-core holding the keystone: the shared ONNX
embedder (one model, loaded once, reused across every corpus). docs-next imports it
(retrieve/guard repointed; lib/rag/embed.ts kept as a thin shim for the other
importers). Behaviour identical. Foundation for the central Ask backend.
