+++ apps/docs-next/package.json
@@ -24,6 +24,7 @@
   },
   "dependencies": {
     "@agentskit/adapters": "workspace:*",
+    "@agentskit/ask-core": "workspace:*",
     "@agentskit/core": "workspace:*",
     "@agentskit/eval": "workspace:*",
     "@agentskit/memory": "workspace:*",
