+++ packages/ask-core/package.json
@@ -1,0 +1,35 @@
+{
+  "name": "@agentskit/ask-core",
+  "version": "0.0.0",
+  "private": true,
+  "description": "Shared engine for the AgentsKit Ask backend (RFC-0007): embedder, guards, streaming protocol, and the framework-neutral createAskHandler. Reused by the docs site and the central Railway backend.",
+  "type": "module",
+  "sideEffects": false,
+  "main": "./dist/index.cjs",
+  "module": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js",
+      "require": "./dist/index.cjs"
+    }
+  },
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build": "tsup",
+    "lint": "tsc --noEmit",
+    "dev": "tsup --watch"
+  },
+  "dependencies": {
+    "@agentskit/core": "workspace:*",
+    "@huggingface/transformers": "^4.2.0"
+  },
+  "devDependencies": {
+    "@types/node": "^25.9.3",
+    "tsup": "^8.5.1",
+    "typescript": "^6.0.3"
+  }
+}
