+++ packages/ask-core/tsup.config.ts
@@ -1,0 +1,12 @@
+import { defineConfig } from 'tsup'
+
+export default defineConfig({
+  entry: { index: 'src/index.ts' },
+  format: ['esm', 'cjs'],
+  dts: { compilerOptions: { ignoreDeprecations: '6.0' } },
+  sourcemap: true,
+  clean: false,
+  treeshake: true,
+  // @huggingface/transformers is heavy + native (onnxruntime) — never bundle it.
+  external: ['@huggingface/transformers', '@agentskit/core'],
+})
